normalize older classes to new APIs.
