--- pythomata/v3/dfas/__init__.py
@@ -1,1 +1,0 @@
-"""This package contains DFAs implementations."""
