+++ setup.cfg
@@ -35,6 +35,9 @@ strict_optional = True
 [mypy-graphviz.*]
 ignore_missing_imports = True
 
+[mypy-sympy.*]
+ignore_missing_imports = True
+
 # Per-module options for tests dir:
 
 [mypy-pytest]
