+++ tox.ini
@@ -24,7 +24,6 @@ setenv =
     PYTHONPATH = {toxinidir}
 deps = pipenv
 commands =
-    pipenv install --dev
     pip install .
     pytest --basetemp={envtmpdir} --doctest-modules \
         pythomata tests/ \
